requirements.txt
Flask==2.0.2
Flask-Caching==1.10.1
flask-paginate==2021.10.29
Pillow==8.4.0
python-poppler==0.2.2
